Add new report type nvi_institution_status
