+++ report-api/src/main/resources/template/nvi-institution-status.sparql
@@ -1,0 +1,33 @@
+PREFIX : <https://nva.sikt.no/ontology/publication#>
+PREFIX xsd: <http://www.w3.org/2001/XMLSchema#>
+
+SELECT DISTINCT
+  ?STATUS_KONTROLLERT
+  ?PUBLIKASJONSFORM
+  ?PUBLISERINGSKANALTYPE
+  ?ISSN
+  ?KVALITETSNIVAKODE
+  ?PERSONLOPENR
+  ?INSTITUSJONSNR
+  ?AVDNR
+  ?UNDAVDNR
+  ?GRUPPENR
+  ?FORFATTERE_TOTALT
+  ?FORFATTERE_INT
+  ?VEKTET
+  ?ETTERNAVN
+  ?ETTERNAVN
+  ?FORNAVN
+  ?PUBLISERINGSKANALNAVN
+  ?SIDE_FRA
+  ?SIDE_TIL
+  ?SIDEANTALL
+  ?VA_TITTEL
+  ?SPRAK
+  ?RAPPORTERT_STATUS
+  ?VEKTINGSTALL
+  ?FAKTORTALL_SAMARBEID
+  ?FORFATTERDEL
+  ?FORFATTERVEKT
+  WHERE {
+  } LIMIT __PAGE_SIZE__ OFFSET __OFFSET__
