+++ report-api/src/test/java/no/sikt/nva/data/report/api/fetch/formatter/ExpectedCsvFormatter.java
@@ -43,7 +43,7 @@ public class ExpectedCsvFormatter {
                 .mapToInt(row -> row[index].length())
                 .map(ExpectedCsvFormatter::addQuoteLength)
                 .max()
-                .orElseThrow(),
+                .orElse(0),
             headers[index].length());
     }
