+++ report-api/src/test/java/no/sikt/nva/data/report/api/fetch/testutils/ValidRequestSource.java
@@ -1,6 +1,5 @@
 package no.sikt.nva.data.report.api.fetch.testutils;
 
-import static com.google.common.net.MediaType.MICROSOFT_EXCEL;
 import static no.sikt.nva.data.report.api.fetch.CustomMediaType.TEXT_CSV;
 import static no.sikt.nva.data.report.api.fetch.CustomMediaType.TEXT_PLAIN;
 import java.util.stream.Stream;
@@ -161,9 +160,20 @@ public class ValidRequestSource implements ArgumentsProvider {
                 )
             ),
             Arguments.of(
-                Named.of("nvi — application/vnd.ms-excel", new TestingRequest(
-                             MICROSOFT_EXCEL.toString(),
-                             "nvi",
+                Named.of("nvi institution report — text/csv", new TestingRequest(
+                             TEXT_CSV.toString(),
+                             "nvi-institution-status",
+                             "2024-01-01",
+                             "1998-01-01",
+                             OFFSET,
+                             PAGE_SIZE
+                         )
+                )
+            ),
+            Arguments.of(
+                Named.of("nvi institution report — text/plain", new TestingRequest(
+                             TEXT_PLAIN.toString(),
+                             "nvi-institution-status",
                              "2024-01-01",
                              "1998-01-01",
                              OFFSET,
