+++ report-api/src/test/java/no/sikt/nva/data/report/api/fetch/testutils/generator/NviHeaders.java
@@ -1,0 +1,15 @@
+package no.sikt.nva.data.report.api.fetch.testutils.generator;
+
+public final class NviHeaders {
+
+    public static final String INSTITUTION_POINTS = "institutionPoints";
+    public static final String POINTS_FOR_AFFILIATION = "pointsForAffiliation";
+    public static final String INSTITUTION_APPROVAL_STATUS = "institutionApprovalStatus";
+    public static final String TOTAL_POINTS = "totalPoints";
+    public static final String PUBLICATION_TYPE_CHANNEL_LEVEL_POINTS = "publicationTypeChannelLevelPoints";
+    public static final String AUTHOR_SHARE_COUNT = "authorShareCount";
+    public static final String INTERNATIONAL_COLLABORATION_FACTOR = "internationalCollaborationFactor";
+    public static final String REPORTED_PERIOD = "reportedPeriod";
+    public static final String GLOBAL_APPROVAL_STATUS = "globalApprovalStatus";
+    public static final String IS_APPLICABLE = "isApplicable";
+}
