+++ report-api/src/test/java/no/sikt/nva/data/report/api/fetch/testutils/generator/NviInstitutionStatusHeaders.java
@@ -1,0 +1,31 @@
+package no.sikt.nva.data.report.api.fetch.testutils.generator;
+
+public final class NviInstitutionStatusHeaders {
+
+    public static String INSTITUTION_APPROVAL_STATUS = "STATUS_KONTROLLERT";//TODO: Check if this is correct
+    public static String PUBLICATION_INSTANCE = "PUBLIKASJONSFORM";
+    public static String PUBLICATION_CHANNEL_TYPE = "PUBLISERINGSKANALTYPE";
+    public static String ISSN = "ISSN";
+    public static String PUBLICATION_CHANNEL_LEVEL = "KVALITETSNIVAKODE";
+    public static String CONTRIBUTOR_IDENTIFIER = "PERSONLOPENR";
+    public static final String INSTITUTION_ID = "INSTITUSJONSNR";
+    public static final String FACULTY_ID = "AVDNR";
+    public static final String DEPARTMENT_ID = "UNDAVDNR";
+    public static final String GROUP_ID = "GRUPPENR";
+    public static final String AUTHOR_COUNT = "FORFATTERE_TOTALT";
+    public static final String AUTHOR_INT = "FORFATTERE_INT";//TODO: Find out what this is
+    public static final String PUBLICATION_TYPE_CHANNEL_LEVEL_POINTS = "VEKTET"; //TODO: Check if this is correct
+    public static final String LAST_NAME = "ETTERNAVN";
+    public static final String FIRST_NAME = "FORNAVN";
+    public static final String PUBLICATION_CHANNEL_NAME = "PUBLISERINGSKANALNAVN";
+    public static final String PAGE_FROM = "SIDE_FRA";
+    public static final String PAGE_TO = "SIDE_TIL";
+    public static final String PAGE_COUNT = "SIDEANTALL";
+    public static final String PUBLICATION_TITLE = "VA_TITTEL";
+    public static final String LANGUAGE = "SPRAK";
+    public static final String IS_REPORTED = "RAPPORTERT_STATUS"; //TODO: Check if this is correct
+    public static final String TOTAL_POINTS = "VEKTINGSTALL"; //TODO: Check if this is correct
+    public static final String INTERNATIONAL_COLLABORATION_FACTOR = "FAKTORTALL_SAMARBEID";
+    public static final String AUTHOR_SHARE_COUNT = "FORFATTERDEL";
+    public static final String POINTS_FOR_AFFILIATION = "FORFATTERVEKT";
+}
